Updating and fixing KDtree implementation
git-svn-id: https://ncwms.svn.sourceforge.net/svnroot/ncwms/trunk@877 f020452a-5e1f-0410-8fa4-ae2b6fb27d5e

+++ src/java/uk/ac/rdg/resc/edal/cdm/kdtree/NonTerminalTreeNode.java
@@ -1,10 +1,10 @@
 package uk.ac.rdg.resc.edal.cdm.kdtree;
 
 public class NonTerminalTreeNode extends TreeNode {
-	float discriminator;
+	double discriminator;
 	boolean is_latitude;
 	
-	public NonTerminalTreeNode(float discriminator, boolean isLatitude) {
+	public NonTerminalTreeNode(double discriminator, boolean isLatitude) {
 		super();
 		this.discriminator = discriminator;
 		is_latitude = isLatitude;
